PCE: Started porting old ExPCE code
Started porting over code from ExPCE (ca. 2010, standalone), still needs loads of reworking and rethinking

+++ MasterFudgeMk2/Media/Coleco/BaseCartridge.cs
@@ -20,7 +20,7 @@ namespace MasterFudgeMk2.Media.Coleco
         public virtual void Reset() { }
         public virtual void Unload() { }
         public virtual void Step() { }
-        public abstract byte Read(ushort address);
-        public abstract void Write(ushort address, byte value);
+        public abstract byte Read(uint address);
+        public abstract void Write(uint address, byte value);
     }
 }
