+++ MasterFudgeMk2/Media/IMedia.cs
@@ -15,7 +15,7 @@ namespace MasterFudgeMk2.Media
         void Reset();
         void Unload();
         void Step();
-        byte Read(ushort address);
-        void Write(ushort address, byte value);
+        byte Read(uint address);
+        void Write(uint address, byte value);
     }
 }
